write handleSubmit fn to pass date ranges up to App.js from DatePicker.js; modify App.js to send dates from DatePicker to make new Api call; tweak styles to create responsive grid layout for apod cards.
I wrote the handleSubmit fn to handle all cases for errors in user input (in which case there's no api call) and all cases for where we can make a successful api call.
Dates are rejected for chronology errors, out-of-range errors, and when both dates are invalid.
If either or both dates are valid and in-range, we make an api call.

I modified App.js so now the url for the api call is reformatted for each new date range
sent from DatePicker.js. Furthermore, the cards are now displayed in reverse
chronological order.