create Hero component and add DatePicker to Hero; move date formatting fns to dateFormatters.js; make today's Astronomy Picture of the Day the Hero background image; switch label and input positions in DatePicker component; add apod logo to Hero; start Hero styles.
I created the Hero section/component so it would
have a dynamic background-image based on what
today's Astronomy Picture of the Day image is.
The DatePicker component/inputs were moved into
the Hero section so the user can choose right
away the APoD date range/images they want to
view. I mostly completed the desktop styles
for the Hero section, which focused a lot on
on the inputs themselves (focus and input styles
so the user knows which input they're on).