create conditionally rendered header (fixed vs static); place another DatePicker in header; start styles for mobile.
I added scroll and resize listeners that
handle adding the fixed header based on
if the user is below the date range
selector at the bottom of Hero section.
I also added another DatePicker component
to the fixed header so the user can select
a new date range no matter where they're at
on the screen. I also added a nice little
fade in animation so the header doesn't just
pop in.